+++ examples/service-worker.js
@@ -6,7 +6,8 @@ self.addEventListener('install', (e) => {
 self.addEventListener('fetch', (e) => {
   const isTarget =
         e.request.url.startsWith('http') &&
-        e.request.url.includes('recording');
+        (e.request.url.includes('recording') ||
+         e.request.url.includes('lyra-benchmark'));
   if(!isTarget) {
     e.respondWith(fetch(e.request));
     return;
@@ -22,9 +23,6 @@ self.addEventListener('fetch', (e) => {
       const headers = new Headers(response.headers);
       headers.set("Cross-Origin-Embedder-Policy", "require-corp");
       headers.set("Cross-Origin-Opener-Policy", "same-origin");
-
-      // Chrome でローカルの http で実行する場合に必要
-      headers.set("Content-Security-Policy", "treat-as-public-address");
 
       return new Response(response.body, {
         status: response.status,