+++ src/lyra.ts
@@ -1,108 +1,69 @@
-import loadLyraWasmModule from "./lyra_wasm.js";
-import * as lyra_wasm from "./lyra_wasm.js";
+import {
+  LYRA_VERSION,
+  DEFAULT_SAMPLE_RATE,
+  DEFAULT_BITRATE,
+  DEFAULT_ENABLE_DTX,
+  DEFAULT_CHANNELS,
+  LyraDecoderOptions,
+  LyraEncoderOptions,
+  SampleRate,
+  NumberOfChannels,
+  Bitrate,
+} from "./utils";
 
-const MEMFS_MODEL_PATH = "/tmp/";
-
-/**
- * Lyra のエンコード形式のバージョン。
- *
- * エンコード形式に非互換な変更が入った時点での google/lyra のバージョンが格納されている。
- */
-const LYRA_VERSION = "1.3.0";
-
-/**
- * {@link LyraModule.createEncoder} メソッドに指定可能なオプション
- */
-interface LyraEncoderOptions {
-  /**
-   * 入力音声データのサンプルレート
-   *
-   * なお 16000 以外のサンプルレートが指定された場合には、内部的にはリサンプルが行われる。
-   *
-   * デフォルト値: 16000
-   */
-  sampleRate?: 8000 | 16000 | 32000 | 48000;
-
-  /**
-   * 入力音声データのチャンネル数
-   *
-   * 現在は 1 (モノラル）のみが指定可能
-   */
-  numberOfChannels?: 1;
-
-  /**
-   * エンコード後の音声データのビットレート
-   *
-   * デフォルト値: 9200
-   */
-  bitrate?: 3200 | 6000 | 9200;
-
-  /**
-   * DTX（discontinuous transmission）を有効にするかどうか
-   *
-   * デフォルト値: false
-   */
-  enableDtx?: boolean;
-}
+const WEB_WORKER_SCRIPT = "__WEB_WORKER_SCRIPT__";
 
 /**
- * {@link LyraModule.createDecoder} メソッドに指定可能なオプション
- */
-interface LyraDecoderOptions {
-  /**
-   * 入力音声データのサンプルレート
-   *
-   * デフォルト値: 16000
-   */
-  sampleRate?: 8000 | 16000 | 32000 | 48000;
-
-  /**
-   * 入力音声データのチャンネル数
-   *
-   * 現在は 1 (モノラル）のみが指定可能
-   */
-  numberOfChannels?: 1;
-}
-
-const DEFAULT_SAMPLE_RATE = 16000;
-const DEFAULT_BITRATE = 9200;
-const DEFAULT_ENABLE_DTX = false;
-const DEFAULT_CHANNELS = 1;
-
-const FRAME_DURATION_MS = 20;
-
-/**
- * Lyra 用の WebAssembly ファイルやモデルファイルを管理するためのクラス
+ * Lyra 用の WebAssembly ファイルやモデルファイルのロードや web worker の管理を行うためのクラス
  */
 class LyraModule {
-  private wasmModule: lyra_wasm.LyraWasmModule;
+  private worker: Worker;
 
-  private constructor(wasmModule: lyra_wasm.LyraWasmModule) {
-    this.wasmModule = wasmModule;
+  private constructor(worker: Worker) {
+    this.worker = worker;
   }
 
   /**
-   * Lyra の WebAssembly ファイルやモデルファイルをロードして {@link LyraModule} のインスタンスを生成する
+   * Lyra の WebAssembly ファイルやモデルファイルをロードや web worker の起動を行い、 {@link LyraModule} のインスタンスを生成する
    *
    * @param wasmPath lyra.wasm および lyra.worker.js が配置されているディレクトリのパスないし URL
    * @param modelPath Lyra 用の *.binarypb および *.tflite が配置されているディレクトリのパスないし URL
    * @returns 生成された {@link LyraModule} インスタンス
    */
-  static async load(wasmPath: string, modelPath: string): Promise<LyraModule> {
-    const wasmModule = await loadLyraWasmModule({
-      locateFile: (path) => {
-        return trimLastSlash(wasmPath) + "/" + path;
-      },
-      preRun: (wasmModule) => {
-        const fileNames = ["lyra_config.binarypb", "soundstream_encoder.tflite", "quantizer.tflite", "lyragan.tflite"];
-        for (const fileName of fileNames) {
-          const url = trimLastSlash(modelPath) + "/" + fileName;
-          wasmModule.FS_createPreloadedFile(MEMFS_MODEL_PATH, fileName, url, true, false);
-        }
-      },
+  static load(wasmPath: string, modelPath: string): Promise<LyraModule> {
+    // lyra は SharedArrayBufffer を使うので COEP / COOP 応答ヘッダ周りの対処が必要になるが、
+    // chrome / firefox と safari で挙動が異なる（前者は COEP / COOP ヘッダが必要で、後者はそれがあるとエラーになる）ので
+    // その問題を回避するために object url で worker を生成するようにする
+    const webWorkerScriptObjectUrl = URL.createObjectURL(
+      new Blob([atob(WEB_WORKER_SCRIPT)], { type: "application/javascript" })
+    );
+    const worker = new Worker(webWorkerScriptObjectUrl, {
+      name: "lyra_sync_worker",
     });
 
-    return new LyraModule(wasmModule);
+    // 各種ファイルは web worker の中でロードされることになるので、
+    // 事前に絶対 URL に変換しておく必要がある
+    wasmPath = new URL(wasmPath, document.location.href).toString();
+    modelPath = new URL(modelPath, document.location.href).toString();
+
+    const promise: Promise<LyraModule> = new Promise((resolve, reject) => {
+      type Response = { data: { type: "LyraModule.load.result"; result: { error?: Error } } };
+      worker.addEventListener(
+        "message",
+        (res: Response) => {
+          const error = res.data.result.error;
+          if (error === undefined) {
+            resolve(new LyraModule(worker));
+          } else {
+            reject(error);
+          }
+        },
+        { once: true }
+      );
+    });
+
+    worker.postMessage({ type: "LyraModule.load", modelPath, wasmPath });
+    return promise;
   }
 
   /**
@@ -110,28 +71,41 @@ class LyraModule {
    *
    * 生成したインスタンスが不要になったら {@link LyraEncoder.destroy} メソッドを呼び出してリソースを解放すること
    *
+   * なお、同じオプションで複数回 {@link createEncoder} メソッドが呼び出された場合には、
+   * 内部的には（wasm レベルでは）同じエンコーダインスタンスが共有して使い回されることになり、
+   * エンコーダ用に確保された wasm メモリ領域は、生成された全ての {@link LyraEncoder} が
+   * {@link LyraEncoder.destroy} を呼び出すまでは解放されない
+   *
    * @params options エンコーダに指定するオプション
    * @returns 生成された {@link LyraEncoder} インスタンス
    */
-  createEncoder(options: LyraEncoderOptions = {}): LyraEncoder {
-    checkSampleRate(options.sampleRate);
-    checkNumberOfChannels(options.numberOfChannels);
-    checkBitrate(options.bitrate);
+  createEncoder(options: LyraEncoderOptions = {}): Promise<LyraEncoder> {
+    const channel = new MessageChannel();
 
-    const encoder = this.wasmModule.LyraEncoder.create(
-      options.sampleRate || DEFAULT_SAMPLE_RATE,
-      options.numberOfChannels || DEFAULT_CHANNELS,
-      options.bitrate || DEFAULT_BITRATE,
-      options.enableDtx || DEFAULT_ENABLE_DTX,
-      MEMFS_MODEL_PATH
-    );
-    if (encoder === undefined) {
-      throw new Error("failed to create lyra encoder");
-    } else {
-      const frameSize = ((options.sampleRate || DEFAULT_SAMPLE_RATE) * FRAME_DURATION_MS) / 1000;
-      const buffer = this.wasmModule.newAudioData(frameSize);
-      return new LyraEncoder(this.wasmModule, encoder, buffer, options);
-    }
+    const promise: Promise<LyraEncoder> = new Promise((resolve, reject) => {
+      type Response = {
+        data: {
+          type: "LyraModule.createEncoder.result";
+          result: { frameSize: number } | { error: Error };
+        };
+      };
+      channel.port1.addEventListener(
+        "message",
+        (res: Response) => {
+          const result = res.data.result;
+          if ("error" in result) {
+            reject(result.error);
+          } else {
+            resolve(new LyraEncoder(channel.port1, result.frameSize, options));
+          }
+        },
+        { once: true }
+      );
+      channel.port1.start();
+    });
+
+    this.worker.postMessage({ type: "LyraModule.createEncoder", port: channel.port2, options }, [channel.port2]);
+    return promise;
   }
 
   /**
@@ -139,49 +113,93 @@ class LyraModule {
    *
    * 生成したインスタンスが不要になったら {@link LyraDecoder.destroy} メソッドを呼び出してリソースを解放すること
    *
+   * なお、同じオプションで複数回 {@link createDecoder} メソッドが呼び出された場合には、
+   * 内部的には（wasm レベルでは）同じデコーダインスタンスが共有して使い回されることになり、
+   * エンコーダ用に確保された wasm メモリ領域は、生成された全ての {@link LyraDecoder} が
+   * {@link LyraDecoder.destroy} を呼び出すまでは解放されない
+   *
    * @params options デコーダに指定するオプション
    * @returns 生成された {@link LyraDecoder} インスタンス
    */
-  createDecoder(options: LyraDecoderOptions = {}): LyraDecoder {
-    checkSampleRate(options.sampleRate);
-    checkNumberOfChannels(options.numberOfChannels);
+  createDecoder(options: LyraDecoderOptions = {}): Promise<LyraDecoder> {
+    const channel = new MessageChannel();
 
-    const decoder = this.wasmModule.LyraDecoder.create(
-      options.sampleRate || DEFAULT_SAMPLE_RATE,
-      options.numberOfChannels || DEFAULT_CHANNELS,
-      MEMFS_MODEL_PATH
-    );
-    if (decoder === undefined) {
-      throw new Error("failed to create lyra decoder");
-    } else {
-      const buffer = this.wasmModule.newBytes();
-      return new LyraDecoder(this.wasmModule, decoder, buffer, options);
-    }
+    const promise: Promise<LyraDecoder> = new Promise((resolve, reject) => {
+      type Response = {
+        data: {
+          type: "LyraModule.createDecoder.result";
+          result: { frameSize: number } | { error: Error };
+        };
+      };
+      channel.port1.addEventListener(
+        "message",
+        (res: Response) => {
+          const result = res.data.result;
+          if ("error" in result) {
+            reject(result.error);
+          } else {
+            resolve(new LyraDecoder(channel.port1, result.frameSize, options));
+          }
+        },
+        { once: true }
+      );
+      channel.port1.start();
+    });
+
+    this.worker.postMessage({ type: "LyraModule.createDecoder", port: channel.port2, options }, [channel.port2]);
+    return promise;
   }
+}
+
+/**
+ * {@link LyraEncoder} の復元に必要な状態
+ *
+ * 詳細は {@link LyraEncoder.fromState()} を参照
+ */
+interface LyraEncoderState {
+  port: MessagePort;
+  sampleRate: SampleRate;
+  numberOfChannels: NumberOfChannels;
+  bitrate: Bitrate;
+  enableDtx: boolean;
+  frameSize: number;
+}
+
+/**
+ * {@link LyraDecoder} の復元に必要な状態
+ *
+ * 詳細は {@link LyraDecoder.fromState()} を参照
+ */
+interface LyraDecoderState {
+  port: MessagePort;
+  sampleRate: SampleRate;
+  numberOfChannels: NumberOfChannels;
+  frameSize: number;
 }
 
 /**
  * Lyra のエンコーダ
  */
 class LyraEncoder {
-  private wasmModule: lyra_wasm.LyraWasmModule;
-  private encoder: lyra_wasm.LyraWasmEncoder;
-  private buffer: lyra_wasm.AudioData;
+  /**
+   * wasm でのエンコード処理を実行する web worker と通信するためのポート
+   */
+  readonly port: MessagePort;
 
   /**
    * 現在のサンププリングレート
    */
-  readonly sampleRate: number;
+  readonly sampleRate: SampleRate;
 
   /**
    * 現在のチャネル数
    */
-  readonly numberOfChannels: number;
+  readonly numberOfChannels: NumberOfChannels;
 
   /**
    * 現在のエンコードビットレート
    */
-  readonly bitrate: number;
+  readonly bitrate: Bitrate;
 
   /**
    * DTX が有効になっているかどうか
@@ -196,25 +214,21 @@ class LyraEncoder {
   /**
    * @internal
    */
-  constructor(
-    wasmModule: lyra_wasm.LyraWasmModule,
-    encoder: lyra_wasm.LyraWasmEncoder,
-    buffer: lyra_wasm.AudioData,
-    options: LyraEncoderOptions
-  ) {
-    this.wasmModule = wasmModule;
-    this.encoder = encoder;
-    this.buffer = buffer;
-
+  constructor(port: MessagePort, frameSize: number, options: LyraEncoderOptions) {
+    this.port = port;
+    this.frameSize = frameSize;
     this.sampleRate = options.sampleRate || DEFAULT_SAMPLE_RATE;
     this.numberOfChannels = options.numberOfChannels || DEFAULT_CHANNELS;
     this.bitrate = options.bitrate || DEFAULT_BITRATE;
     this.enableDtx = options.enableDtx || DEFAULT_ENABLE_DTX;
-    this.frameSize = buffer.size();
   }
 
   /**
    * 20ms 分の音声データをエンコードする
+   *
+   * なお audioData の所有権は web worker に移転されるので、
+   * このメソッド呼び出し後には呼び出しもとスレッドからはデータに参照できなくなります
+   * （つまり同じ audioData インスタンスの使い回しはできない）
    *
    * @params audioData エンコード対象の音声データ
    * @returns エンコード後のバイト列。もし DTX が有効で音声データが無音な場合には undefined が代わりに返される。
@@ -225,56 +239,55 @@ class LyraEncoder {
    * - 入力音声データが 20ms 単位（サンプル数としては {@link LyraEncoder.frameSize}）ではない
    * - その他、何らかの理由でエンコードに失敗した場合
    */
-  encode(audioData: Int16Array): Uint8Array | undefined {
-    if (audioData.length !== this.frameSize) {
-      throw new Error(
-        `expected an audio data with ${this.frameSize} samples, but got one with ${audioData.length} samples`
+  encode(audioData: Int16Array): Promise<Uint8Array | undefined> {
+    const promise: Promise<Uint8Array | undefined> = new Promise((resolve, reject) => {
+      type Response = {
+        data: {
+          type: "LyraEncoder.encode.result";
+          result: { encodedAudioData: Uint8Array | undefined } | { error: Error };
+        };
+      };
+      this.port.addEventListener(
+        "message",
+        (res: Response) => {
+          const result = res.data.result;
+          if ("error" in result) {
+            reject(result.error);
+          } else {
+            resolve(result.encodedAudioData);
+          }
+        },
+        { once: true }
       );
-    }
+    });
 
-    this.wasmModule.copyInt16ArrayToAudioData(this.buffer, audioData);
-
-    const result = this.encoder.encode(this.buffer);
-
-    if (result === undefined) {
-      throw new Error("failed to encode");
-    } else {
-      try {
-        const encodedAudioData = new Uint8Array(result.size());
-        for (let i = 0; i < encodedAudioData.length; i++) {
-          encodedAudioData[i] = result.get(i);
-        }
-
-        if (encodedAudioData.length === 0) {
-          // DTX が有効、かつ、 audioData が無音ないしノイズだけを含んでいる場合にはここに来る
-          return undefined;
-        }
-        return encodedAudioData;
-      } finally {
-        result.delete();
-      }
-    }
-  }
-
-  /**
-   * エンコードビットレートを変更する
-   *
-   * @params bitrate 変更後のビットレート
-   */
-  setBitrate(bitrate: number): void {
-    checkBitrate(bitrate);
-
-    if (!this.encoder.setBitrate(bitrate)) {
-      throw new Error(`failed to update bitrate from ${this.bitrate} to ${bitrate}`);
-    }
+    this.port.postMessage({ type: "LyraEncoder.encode", audioData }, [audioData.buffer]);
+    return promise;
   }
 
   /**
    * エンコーダ用に確保したリソースを解放する
    */
   destroy(): void {
-    this.encoder.delete();
-    this.buffer.delete();
+    this.port.postMessage({ type: "LyraEncoder.destroy" });
+    this.port.close();
+  }
+
+  /**
+   * {@link LyraEncoderState} から {@link LyraEncoder} を復元する
+   *
+   * {@link LyraEncoder} は {@link MessagePort.postMessage()} を使って、
+   * 別の web worker に転送することが可能。
+   * ただし、転送時にはクラスやメソッドの状態は落ちてしまうので、
+   * それを復元して再び利用可能にするための関数。
+   * なお、転送の際には {@link LyraEncoder.port} の所有権を移譲する必要がある。
+   *
+   * @param state エンコーダの状態
+   * @return 復元されたエンコーダ
+   */
+  static fromState(state: LyraEncoderState): LyraEncoder {
+    state.port.start();
+    return new LyraEncoder(state.port, state.frameSize, state);
   }
 }
 
@@ -282,19 +295,20 @@ class LyraEncoder {
  * Lyra のデコーダ
  */
 class LyraDecoder {
-  private wasmModule: lyra_wasm.LyraWasmModule;
-  private decoder: lyra_wasm.LyraWasmDecoder;
-  private buffer: lyra_wasm.Bytes;
+  /**
+   * wasm でのデコード処理を実行する web worker と通信するためのポート
+   */
+  readonly port: MessagePort;
 
   /**
    * 現在のサンププリングレート
    */
-  readonly sampleRate: number;
+  readonly sampleRate: SampleRate;
 
   /**
    * 現在のチャネル数
    */
-  readonly numberOfChannels: number;
+  readonly numberOfChannels: NumberOfChannels;
 
   /**
    * 一つのフレーム（{@link LyraEncoder.decode} メソッドの返り値の音声データ）に含まれるサンプル数
@@ -304,99 +318,89 @@ class LyraDecoder {
   /**
    * @internal
    */
-  constructor(
-    wasmModule: lyra_wasm.LyraWasmModule,
-    decoder: lyra_wasm.LyraWasmDecoder,
-    buffer: lyra_wasm.Bytes,
-    options: LyraDecoderOptions
-  ) {
-    this.wasmModule = wasmModule;
-    this.decoder = decoder;
-    this.buffer = buffer;
-
+  constructor(port: MessagePort, frameSize: number, options: LyraDecoderOptions) {
+    this.port = port;
+    this.frameSize = frameSize;
     this.sampleRate = options.sampleRate || DEFAULT_SAMPLE_RATE;
     this.numberOfChannels = options.numberOfChannels || DEFAULT_CHANNELS;
-
-    this.frameSize = (this.sampleRate * FRAME_DURATION_MS) / 1000;
   }
 
   /**
    * {@link LyraEncoder.encode} メソッドによってエンコードされた音声データをデコードする
    *
+   * なお encodedAudioData の所有権は web worker に移転されるので、
+   * このメソッド呼び出し後には呼び出しもとスレッドからはデータに参照できなくなります
+   * （つまり同じ encodedAudioData インスタンスの使い回しはできない）
+   *
    * @params encodedAudioData デコード対象のバイナリ列ないし undefined
    * @returns デコードされた 20ms 分の音声データ。undefined が渡された場合には代わりにコンフォートノイズが生成される。
    */
-  decode(encodedAudioData: Uint8Array | undefined): Int16Array {
-    if (encodedAudioData !== undefined) {
-      this.buffer.resize(0, 0); // clear() を使うと「関数が存在しない」というエラーが出るので resize() で代用
-      for (const v of encodedAudioData) {
-        this.buffer.push_back(v);
-      }
-      if (!this.decoder.setEncodedPacket(this.buffer)) {
-        throw new Error("failed to set encoded packet");
-      }
-    }
+  decode(encodedAudioData: Uint8Array | undefined): Promise<Int16Array> {
+    const promise: Promise<Int16Array> = new Promise((resolve, reject) => {
+      type Response = {
+        data: {
+          type: "LyraDecoder.decode.result";
+          result: { audioData: Int16Array } | { error: Error };
+        };
+      };
+      this.port.addEventListener(
+        "message",
+        (res: Response) => {
+          const result = res.data.result;
+          if ("error" in result) {
+            reject(result.error);
+          } else {
+            resolve(result.audioData);
+          }
+        },
+        { once: true }
+      );
+    });
 
-    const result = this.decoder.decodeSamples(this.frameSize);
-
-    if (result === undefined) {
-      throw Error("failed to decode samples");
+    if (encodedAudioData === undefined) {
+      this.port.postMessage({ type: "LyraDecoder.decode", encodedAudioData });
+    } else {
+      this.port.postMessage({ type: "LyraDecoder.decode", encodedAudioData }, [encodedAudioData.buffer]);
     }
-    try {
-      const audioData = new Int16Array(this.frameSize);
-      this.wasmModule.copyAudioDataToInt16Array(audioData, result);
-      return audioData;
-    } finally {
-      result.delete();
-    }
+    return promise;
   }
 
   /**
    * デコーダ用に確保したリソースを解放する
    */
   destroy(): void {
-    this.decoder.delete();
-    this.buffer.delete();
+    this.port.postMessage({ type: "LyraDecoder.destroy" });
+    this.port.close();
+  }
+
+  /**
+   * {@link LyraDecoderState} から {@link LyraDecoder} を復元する
+   *
+   * {@link LyraDecoder} は {@link MessagePort.postMessage()} を使って、
+   * 別の web worker に転送することが可能。
+   * ただし、転送時にはクラスやメソッドの状態は落ちてしまうので、
+   * それを復元して再び利用可能にするための関数。
+   * なお、転送の際には {@link LyraDecoder.port} の所有権を移譲する必要がある。
+   *
+   * @param state デコーダの状態
+   * @return 復元されたデコーダ
+   */
+  static fromState(state: LyraDecoderState): LyraDecoder {
+    state.port.start();
+    return new LyraDecoder(state.port, state.frameSize, state);
   }
 }
 
-function trimLastSlash(s: string): string {
-  if (s.slice(-1) === "/") {
-    return s.slice(0, -1);
-  }
-  return s;
-}
-
-function checkSampleRate(n: number | undefined): void {
-  switch (n) {
-    case undefined:
-    case 8000:
-    case 16000:
-    case 32000:
-    case 48000:
-      return;
-  }
-  throw new Error(`unsupported sample rate: expected one of 8000, 16000, 32000 or 48000, but got ${n}`);
-}
-
-function checkNumberOfChannels(n: number | undefined): void {
-  switch (n) {
-    case undefined:
-    case 1:
-      return;
-  }
-  throw new Error(`unsupported number of channels: expected 1, but got ${n}`);
-}
-
-function checkBitrate(n: number | undefined): void {
-  switch (n) {
-    case undefined:
-    case 3200:
-    case 6000:
-    case 9200:
-      return;
-  }
-  throw new Error(`unsupported bitrate: expected one of 3200, 6000 or 9200, but got ${n}`);
-}
-
-export { LyraModule, LyraDecoder, LyraEncoder, LyraEncoderOptions, LyraDecoderOptions, LYRA_VERSION };
+export {
+  LYRA_VERSION,
+  LyraDecoderOptions,
+  LyraEncoderOptions,
+  LyraModule,
+  LyraEncoder,
+  LyraDecoder,
+  LyraEncoderState,
+  LyraDecoderState,
+  SampleRate,
+  NumberOfChannels,
+  Bitrate,
+};